chatbot_service/requirements.txt
google-generativeai==0.8.5
Django==5.2.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
djangorestframework==3.16.0

